fix: disable changing modes when in the process of doing so
