+++ cellacdc/apps.py
@@ -81,6 +81,7 @@ from . import types
 from . import plot
 from . import urls
 from .acdc_regex import float_regex
+from . import _base_widgets
 
 POSITIVE_FLOAT_REGEX = float_regex(allow_negative=False)
 PRE_PROCESSING_STEPS = [
@@ -146,34 +147,9 @@ def addCustomModelMessages(QParent=None):
     
     return modelFilePath
 
-class QBaseDialog(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-    def exec_(self, resizeWidthFactor=None):
-        if resizeWidthFactor is not None:
-            self.show()
-            self.resize(int(self.width()*resizeWidthFactor), self.height())
-        self.show(block=True)
-
-    def show(self, block=False):
-        self.setWindowFlags(Qt.Window | Qt.WindowStaysOnTopHint)
-        super().show()
-        if block:
-            self.loop = QEventLoop()
-            self.loop.exec_()
-
-    def closeEvent(self, event):
-        if hasattr(self, 'loop'):
-            self.loop.exit()
-    
-    def keyPressEvent(self, event) -> None:
-        if event.key() == Qt.Key_Escape:
-            event.ignore()
-            return
-            
-        super().keyPressEvent(event)
-
+class QBaseDialog(_base_widgets.QBaseDialog):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 class customAnnotationDialog(QDialog):
     sigDeleteSelecAnnot = Signal(object)
